move ledger/virtual channel to separate files
