+++ channel/channel.go
@@ -29,132 +29,6 @@ type Channel struct {
 	// Longer term, we should have a more efficient and smart mechanism to store states https://github.com/statechannels/go-nitro/issues/106
 }
 
-type SingleHopVirtualChannel struct {
-	Channel
-}
-
-// NewSingleHopVirtualChannel returns a new SingleHopVirtualChannel based on the supplied state.
-func NewSingleHopVirtualChannel(s state.State, myIndex uint) (*SingleHopVirtualChannel, error) {
-	if myIndex > 2 {
-		return &SingleHopVirtualChannel{}, errors.New("myIndex in a single hop virtual channel must be 0, 1, or 2")
-	}
-	if len(s.Participants) != 3 {
-		return &SingleHopVirtualChannel{}, errors.New("a single hop virtual channel must have exactly three participants")
-	}
-	for _, assetExit := range s.Outcome {
-		if len(assetExit.Allocations) != 2 {
-			return &SingleHopVirtualChannel{}, errors.New("a single hop virtual channel's initial state should only have two allocations")
-		}
-	}
-	c, err := New(s, myIndex)
-
-	return &SingleHopVirtualChannel{*c}, err
-}
-
-// amountAtIndex gets allocations at the specified index and returns the amount.
-func (v SingleHopVirtualChannel) amountAtIndex(index uint) types.Funds {
-	supported, err := v.LatestSupportedState()
-
-	// If there is no supported state we just return an empty amount
-	if err != nil {
-		return types.Funds{}
-	}
-
-	amount := types.Funds{}
-
-	for _, assetExit := range supported.Outcome {
-		asset := assetExit.Asset
-		allocations := assetExit.Allocations
-
-		if index < uint(len(allocations)) {
-			amount[asset] = allocations[index].Amount
-		}
-	}
-	return amount
-}
-
-// LeftAmount returns the amount of the first allocation, which allocates to the left.
-func (v SingleHopVirtualChannel) LeftAmount() types.Funds {
-	return v.amountAtIndex(0)
-}
-
-// RightAmount returns the amount of the second allocation, which allocates to the right.
-func (v SingleHopVirtualChannel) RightAmount() types.Funds {
-	return v.amountAtIndex(1)
-}
-
-// Equal returns true if the supplied SingleHopVirtualChannel is deeply equal to the receiver, false otherwise.
-func (v *SingleHopVirtualChannel) Equal(w *SingleHopVirtualChannel) bool {
-	return v.Channel.Equal(w.Channel)
-}
-
-// Clone returns a pointer to a new, deep copy of the receiver, or a nil pointer if the receiver is nil.
-func (v *SingleHopVirtualChannel) Clone() *SingleHopVirtualChannel {
-	if v == nil {
-		return nil
-	}
-	w := SingleHopVirtualChannel{*v.Channel.Clone()}
-	return &w
-}
-
-const proposerIndex = uint(0)
-
-type TwoPartyLedger struct {
-	Channel
-}
-
-// NewTwoPartyLedger constructs a new two party ledger channel from the supplied state.
-func NewTwoPartyLedger(s state.State, myIndex uint) (*TwoPartyLedger, error) {
-	if myIndex > 1 {
-		return &TwoPartyLedger{}, errors.New("myIndex in a two party ledger channel must be 0 or 1")
-	}
-	if len(s.Participants) != 2 {
-		return &TwoPartyLedger{}, errors.New("two party ledger channels must have exactly two participants")
-	}
-
-	c, err := New(s, myIndex)
-
-	return &TwoPartyLedger{*c}, err
-}
-
-// Equal returns true if the supplied TwoPartyLedger is deeply equal to the receiver, false otherwise.
-func (lc *TwoPartyLedger) Equal(lc2 *TwoPartyLedger) bool {
-	return lc.Channel.Equal(lc2.Channel)
-}
-
-// Clone returns a pointer to a new, deep copy of the receiver, or a nil pointer if the receiver is nil.
-func (lc *TwoPartyLedger) Clone() *TwoPartyLedger {
-	if lc == nil {
-		return nil
-	}
-	w := TwoPartyLedger{*lc.Channel.Clone()}
-	return &w
-}
-
-// Proposed returns the latest proposed ledger state that is only signed by the proposer.
-// If the latest state signed by the proposer is the supported state then Proposed returns a nil state and false.
-func (lc *TwoPartyLedger) Proposed() (state.State, bool) {
-
-	highestSignedByProposer := uint64(0)
-
-	for turnNum, signedState := range lc.SignedStateForTurnNum {
-		if signedByProposer := signedState.HasSignatureForParticipant(proposerIndex); signedByProposer && turnNum > highestSignedByProposer {
-			highestSignedByProposer = turnNum
-		}
-	}
-
-	if highestSignedByProposer == lc.latestSupportedStateTurnNum {
-		return state.State{}, false
-	} else {
-		return lc.SignedStateForTurnNum[highestSignedByProposer].State(), true
-	}
-}
-
-// IsProposer returns true if the we are responsible for proposing ledger updates, false otherwise
-func (lc *TwoPartyLedger) IsProposer() bool {
-	return lc.MyIndex == proposerIndex
-}
-
 // New constructs a new Channel from the supplied state.
 func New(s state.State, myIndex uint) (*Channel, error) {
 	c := Channel{}
